Refactor: move UserRepository forom domain package to infra package
