+++ src/main/java/com/devthink/devthink_server/application/AuthenticationService.java
@@ -2,7 +2,7 @@ package com.devthink.devthink_server.application;
 
 
 import com.devthink.devthink_server.domain.User;
-import com.devthink.devthink_server.domain.UserRepository;
+import com.devthink.devthink_server.infra.UserRepository;
 import com.devthink.devthink_server.errors.InvalidTokenException;
 import com.devthink.devthink_server.errors.LoginFailException;
 import com.devthink.devthink_server.utils.JwtUtil;
