+++ src/main/java/com/devthink/devthink_server/infra/UserRepository.java
@@ -1,5 +1,6 @@
-package com.devthink.devthink_server.domain;
+package com.devthink.devthink_server.infra;
 
+import com.devthink.devthink_server.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
